% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_neighbor_expression.R
\name{extract_neighborexpr_by_gene}
\alias{extract_neighborexpr_by_gene}
\title{Extract neighboring cell type expression for a particular gene}
\usage{
extract_neighborexpr_by_gene(
  neighbor_expr_list,
  gene,
  expr_list_names = NULL,
  cell_IDs = NULL
)
}
\arguments{
\item{neighbor_expr_list}{list of matrices by cell type group from \code{measure_neighbor_expression_by_celltype.}}

\item{gene}{gene to make data.table for.}
}
\description{
Extract neighboring cell type expression for a particular gene from an
existing object produced by \code{smiDE::measure_neighbor_expression_by_celltype}.
Columns of this table can be used as covariates in DE, specified in \code{formula} argument of \code{run_de} or \code{smi_de}.
}
\examples{
library(Seurat)
library(data.table); setDTthreads(1)
datadir<-system.file("extdata", package="smiDE")
sem <- readRDS(paste0(datadir, "/small_nsclc.rds"))
metainfo <- data.table(sem@meta.data)

cell_adjacency_dt <- 
  fast_make_all_neighbors(dt = metainfo
                          ,cellid_col = "cell_ID"
                          ,sdimx_col = "sdimx"
                          ,sdimy_col = "sdimy"
                          ,radius = 0.05 
  )

cell_adjacency_dt[,.(from,to,distance,weight,cell_type_from, cell_type_to)]

totalcounts <- sem@meta.data[["totalcounts"]]
sem <- Seurat::SetAssayData(sem
                           ,"data"
                           ,sem[["RNA"]]@counts \%*\% Matrix::Diagonal(x=mean(totalcounts)/totalcounts, names=colnames(sem))
                           )
                           
## weighted by distance.  Using normalized counts
nblist <- 
  measure_neighbor_expression_by_celltype(assay_matrix =  sem[["RNA"]]@data
                                          ,metadata=metainfo
                                          ,cell_adjacencies = cell_adjacency_dt
                                          ,cell_type_metadata_colname = "cell_type"
                                          ,ref_celltype = "mast"
                                          ,weight_colname = "weight"
  )

spp1_neighbor_expr <- 
  extract_neighborexpr_by_gene(nblist$neighbor_expr_byct, gene= "SPP1")


## unweighted, using raw counts
nblist_unweighted_counts <- 
  measure_neighbor_expression_by_celltype(assay_matrix = sem[["RNA"]]@counts
                                          ,metadata=metainfo
                                          ,cell_adjacencies = cell_adjacency_dt
                                          ,cell_type_metadata_colname = "cell_type"
                                          ,ref_celltype = "mast"
                                          ,weight_colname = NULL
  )

spp1_neighbor_expr_cts <- 
  extract_neighborexpr_by_gene(nblist_unweighted_counts$neighbor_expr_byct, gene= "SPP1")

## Note that the sum of expression for individual cell types is equal to total 'all cell type' / 'allct' expression.
## Note that the 'otherct' column indicates expression from cell types other than the reference cell type (in this case "mast").
spp1_neighbor_expr_cts[1:3]

}
