% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_neighbor_expression.R
\name{measure_neighbor_expression_by_celltype}
\alias{measure_neighbor_expression_by_celltype}
\title{Measure gene expression in neighboring cells}
\usage{
measure_neighbor_expression_by_celltype(
  assay_matrix,
  metadata,
  cell_adjacencies,
  ref_celltype,
  cell_type_metadata_colname,
  aggregation = "sum",
  cellid_colname = "cell_ID",
  weight_colname = NULL,
  Wmat = NULL,
  ct_matrix = NULL,
  adjacency_counts_by_ct = NULL,
  nb_celltypes_to_individually_calc = NULL
)
}
\arguments{
\item{assay_matrix}{(normalized) counts matrix.  Normalized is recommended when output of this function is passed to downstream DE to adjust for neighboring cell expression.}

\item{metadata}{data.table or data.frame of meta.data associated with assay.
Should contain the cell_type_metadata_colname, and cellid_colname.}

\item{cell_adjacencies}{data.table of cell connections with columns 'from' and 'to', such as that in Giotto spatial netowork DT object.}

\item{ref_celltype}{calculate neighbor expression for cells of this particular \code{ref_celltype}.}

\item{cell_type_metadata_colname}{column name in meta.data of cell types}

\item{aggregation}{"sum" (default), calculates total (weighted) sum amount of neighbor expression.  Otherwise if "average" ,a (weighted) average.}

\item{cellid_colname}{column name  of cell ids. Or if missing,}

\item{weight_colname}{If non-missing, uses the "weight_colname" column of cell_adjacencies to weight the calculated expression of neighbors.}

\item{Wmat}{pre-computed, possibly weighted neighborhood adjacency matrix, calculated based on cell_adjacencies data.table.
If missing, will be computed in this function call.}

\item{ct_matrix}{pre-computed sparse matrix indicating cell type for each cell.
If missing, will be computed in this function call.}

\item{adjacency_counts_by_ct}{pre-computed, possibly distance weighted count of neighboring cells by each unique celltype.
If missing, will be computed in this function call.}

\item{nb_celltypes_to_individually_calc}{vector of cell types for which to individually calculate neighbor expression.
default is NULL, which will make a separate list in neighbor_expr_by_ct output with matrix of expression by cell type}
}
\description{
Measure gene expression in neighboring cells using data.table of cell adjacencies.
}
\examples{
library(Seurat)
library(data.table); setDTthreads(1)
datadir<-system.file("extdata", package="smiDE")
sem <- readRDS(paste0(datadir, "/small_nsclc.rds"))
metainfo <- data.table::data.table(sem@meta.data)


tiss <- split(metainfo, by="tissue")

## identify adjacent cells within 0.05 distance of each other within common tissue 
cell_adjacency_dt <- 
  rbindlist(
    lapply(1:length(tiss), function(x){
      nb <- fast_make_all_neighbors(tiss[[x]][,.(cell_ID, sdimx, sdimy,cell_type)]
                                    ,cellid_col = "cell_ID"
                                    ,sdimx_col ="sdimx"
                                    ,sdimy_col ="sdimy"
                                    ,radius = 0.05) 
      return(nb) 
    })
  )

cell_adjacency_dt[]

## unweighted, using raw counts
nblist_unweighted_counts <- 
  measure_neighbor_expression_by_celltype(assay_matrix = sem[["RNA"]]@counts 
                                          ,metadata=metainfo
                                          ,cell_adjacencies = cell_adjacency_dt
                                          ,cell_type_metadata_colname = "cell_type"
                                          ,ref_celltype = "mast"
                                          ,weight_colname = NULL
  )

spp1_neighbor_expr_cts <- 
  extract_neighborexpr_by_gene(nblist_unweighted_counts$neighbor_expr_byct, gene= "SPP1")

## Note that the sum of expression for individual cell types is equal to total 'all cell type' / 'allct' expression.
## Note that the 'otherct' column indicates expression from cell types other than the reference cell type (in this case "mast").
spp1_neighbor_expr_cts[1:3]

totalcounts <- sem@meta.data[["totalcounts"]]
sem <- Seurat::SetAssayData(sem
                           ,"data"
                           ,sem[["RNA"]]@counts \%*\% Matrix::Diagonal(x=mean(totalcounts)/totalcounts, names=colnames(sem))
                           )

## weighted by distance.  Using normalized counts
nblist <- 
  measure_neighbor_expression_by_celltype(assay_matrix = sem[["RNA"]]@data 
                                          ,metadata=metainfo
                                          ,cell_adjacencies = cell_adjacency_dt
                                          ,cell_type_metadata_colname = "cell_type"
                                          ,ref_celltype = "mast"
                                          ,weight_colname = "weight"
  )

names(nblist)

spp1_neighbor_expr <- 
  extract_neighborexpr_by_gene(nblist$neighbor_expr_byct, gene= "SPP1")

spp1_neighbor_expr[1:3]



}
