% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_neighbor_expression.R
\name{measure_neighbor_expression_single}
\alias{measure_neighbor_expression_single}
\title{Measure one gene's expression in neighboring cells, on-the-fly in DE}
\usage{
measure_neighbor_expression_single(
  assay_matrix,
  metadata,
  cell_adjacencies,
  ref_celltype,
  cell_type_metadata_colname,
  aggregation = "sum",
  cellid_colname = "cell_ID",
  weight_colname = NULL,
  Wmat = NULL,
  adjacency_counts_by_ct = NULL,
  nb_celltypes_to_individually_calc = NULL
)
}
\description{
Measure one gene's expression in neighboring cells, on-the-fly in DE
}
