% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_de.R
\name{pre_de}
\alias{pre_de}
\title{Calculate cell adjacencies and neighboring cell expression of genes before DE analysis.}
\usage{
pre_de(
  adjacencies_only = TRUE,
  metadata,
  ref_celltype,
  cell_type_metadata_colname,
  weight_colname = NULL,
  cellid_colname = "cell_ID",
  sdimx_colname = "sdimx",
  sdimy_colname = "sdimy",
  split_neighbors_by_colname = "tissue",
  mm_radius = 0.05,
  counts = NULL,
  normalized_data = NULL,
  verbose = TRUE,
  aggregation = "sum",
  nb_celltypes_to_individually_calc = NULL
)
}
\arguments{
\item{adjacencies_only}{avoid pre-computing expression matrices, only pre-compute the adjacencies between cells.}

\item{metadata}{cell-level metadata corresponding to the cells in counts, with one row per cell.
Should contain the columns of cell_type_metadata_colname,
cellid_colname, split_neighbors_by_colname, sdimx_colname, and sdimy_colname.}

\item{ref_celltype}{Reference category (commonly a cell type), used to calculate neighbor expression with respect to.
Can be vector of one or more levels of cell_type_metadata_colname.
Can use keyword "all" for all levels of cell_type_metadata_colname.
Only relevant if adjacencies_only = FALSE}

\item{cell_type_metadata_colname}{column (commonly cell type column) corresponding to categories by which cell expression is
aggregated by to calculate expression among neighboring cells by cell_type_metadata_colname type.}

\item{weight_colname}{(optional).  If NULL, neighbor expression is unweighted.
If "weight", the column corresponding to "weight" (1/distance) by default is used to weight calculated expression of neighbor cells.
Only relevant if adjacencies_only = FALSE}

\item{cellid_colname}{column name corresponding to cell id.}

\item{sdimx_colname}{column name corresponding to 'x' axis spatial coordinate of cell.}

\item{sdimy_colname}{column name corresponding to 'y' axis spatial coordinate of cell.}

\item{split_neighbors_by_colname}{(optional) If specified, identification of neighboring cells is split by this
column in the meta data.  For example, to avoid identifying cells as neighbors if they were to have close x/y coordinates, but were on different tissue.}

\item{mm_radius}{maximum euclidean distance in units of sdimx, sdimy by which two cells will be identified as neighbors.}

\item{counts}{matrix of counts, only required if adjacencies_only = FALSE}

\item{normalized_data}{optional, matrix of normalized data.
If not provided, totalCounts normalization is applied to counts. Only relevant if adjacencies_only = FALSE}

\item{verbose}{print some extra messages during calculation.}

\item{aggregation}{"sum" (default) calculates a total amount of target expressed in neighboring cells.  Anything other values passed will calculate an average.
Only relevant if adjacencies_only = FALSE}

\item{nb_celltypes_to_individually_calc}{vector of cell types for which to individually calculate neighbor expression.
default is NULL, which will make a separate list in neighbor_expr_by_ct output with matrix of expression by cell type.
Only relevant if adjacencies_only = FALSE}
}
\value{
list object with data.table of cell_adjacencies (one row per pair of adjacent cells),
a list (nblist) with sparse matrices corresponding to neighbor cell expression passed to downstream DE in smi_de or run_de functions.
}
\description{
Calculate cell adjacencies and neighboring cell expression of genes before DE analysis.
}
\examples{
library(Seurat)
library(data.table); setDTthreads(1)
datadir<-system.file("extdata", package="smiDE")
sem <- readRDS(paste0(datadir, "/small_nsclc.rds"))
metainfo <- data.table::data.table(sem@meta.data)
 
pre_de_obj <- 
   pre_de(metadata = metainfo
          ,adjacencies_only = TRUE
          ,cell_type_metadata_colname = "cell_type"
          ,split_neighbors_by_colname = "tissue"
          ,mm_radius = 0.05
          ,sdimx_colname = "sdimx"
          ,sdimy_colname = "sdimy"
          ,verbose=TRUE
   ) 

# data.table of cell-cell adjacencies
pre_de_obj$cell_adjacency_dt[]

### Note:  
### The examples below create genes x cells neighbor-expression matrices.
### This method for pre-calculating the large neighbor expression matrices
### is high-memory and currently deprecated, in favor of doing these calculations
### 'on-the-fly' when running smi_de.
### However, the examples below will still run, and are useful for visual inspection
### of what's being calculated 'under-the-hood' when running DE models and using
### covariate-adjustment to control for neighbor expression.

totalcounts <- sem@meta.data[["totalcounts"]]
sem <- Seurat::SetAssayData(sem
                           ,"data"
                           ,sem[["RNA"]]@counts \%*\% Matrix::Diagonal(x=mean(totalcounts)/totalcounts, names=colnames(sem))
                           )

pre_de_obj <- 
pre_de(adjacencies_only = FALSE
       ,counts = sem[["RNA"]]@counts
       ,normalized_data = sem[["RNA"]]@data
       ,metadata = metainfo
       ,cell_type_metadata_colname = "cell_type"
       ,split_neighbors_by_colname = "tissue"
       ,mm_radius = 0.05
       ,ref_celltype = "fibroblast"
       ,sdimx_colname = "sdimx"
       ,sdimy_colname = "sdimy"
       ,weight_colname = "weight"
       ,aggregation = "sum"
       ,verbose=TRUE
 )
names(pre_de_obj)

# Example for measuring neighbor expression with 'all' levels of
# ref_celltype = 'all'
pre_de_obj <- 
pre_de(counts = sem[["RNA"]]@counts
       ,normalized_data = sem[["RNA"]]@data
       ,metadata = metainfo
       ,cell_type_metadata_colname = "cell_type"
       ,split_neighbors_by_colname = "tissue"
       ,mm_radius = 0.05
       ,ref_celltype = c("all")
       ,sdimx_colname = "sdimx"
       ,sdimy_colname = "sdimy"
       ,weight_colname = "weight"
       ,aggregation = "sum"
       ,verbose=TRUE
)
names(pre_de_obj)
dim(pre_de_obj$nblist$neighbor_expr_byct$otherct)  
metainfo[, .N]


# Example for measuring neighbor expression with a vector of multiple cell types
pre_de_obj <- 
pre_de(counts = sem[["RNA"]]@counts
       ,normalized_data = sem[["RNA"]]@data
       ,metadata = metainfo
       ,cell_type_metadata_colname = "cell_type"
       ,split_neighbors_by_colname = "tissue"
       ,mm_radius = 0.05
       ,ref_celltype = c("macrophage", "fibroblast")
       ,sdimx_colname = "sdimx"
       ,sdimy_colname = "sdimy"
       ,weight_colname = "weight"
       ,aggregation = "sum"
       ,verbose=TRUE
)
names(pre_de_obj)
dim(pre_de_obj$nblist$neighbor_expr_byct$otherct)  
metainfo[cell_type \%in\% pre_de_obj$nblist$ref_celltype, .N]
 
}
