% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{saveVolcano}
\alias{saveVolcano}
\title{Save volcano plot files from list}
\usage{
saveVolcano(volcanoPlot, outFolder)
}
\arguments{
\item{volcanoPlot}{list of volcano plot figures}

\item{outFolder}{output directory}
}
\description{
Save volcano plot files from list
}
\examples{
\dontrun{

datadir <- system.file("extdata", package="smiDE")
de_results <- readRDS(paste0(datadir, "/sample_de_results.rds"))

dat <- results(de_results, comparisons="one.vs.rest", variable="group")[["one.vs.rest"]]
head(dat)

plist <- volcano(de_results, "one.vs.rest", "group", interactive = TRUE)
# tempDir <- tempdir()
# saveVolcano(volcanoPlot = plist, outFolder = getwd())

}

}
